Fix all build errors: API migration, namespaces, and dependencies
CRITICAL FIXES:
- Add missing g_character_system global variable declaration (main.cpp:407)
- Make CharacterSystem::LegacyToVersionedEntityID() public for cross-system use

API MIGRATIONS (CharacterWindow.cpp):
- Component namespaces: game::components:: → game::character::
- Relationships: friends/rivals vectors → GetFriends()/relationships map
- Education: GetFocus()/GetProgress() → direct member access + GetEducationFocusString()
- Life events: GetEvents() → direct life_events member access
- Traits: GetTraitIDs() → active_traits vector iteration
- Family: father_id/mother_id → father/mother

API MIGRATIONS (Influence systems):
- Marriage: spouse_id → spouse (InfluenceCalculator.cpp, InfluenceSystem.cpp)
- Relationships: friends vector → GetFriends() method

NAMESPACE FIXES:
- Logger calls: core::logging::Logger:: → ::core::logging::Logger::
- EntityID type conversions: Use .id member for legacy components

TEMPLATE FIXES:
- std::min ambiguity: Add static_cast<size_t>() for Json::ArrayIndex
- Files: PopulationComponentsSerialization.cpp, CharacterLifeEvents.cpp,
  CharacterRelationships.cpp (6 occurrences)

DEPENDENCY FIXES:
- Make zlib optional with HAVE_ZLIB compile guard
- Add fallback for compression when zlib not available
- Files: CMakeLists.txt, SerializationUtils.cpp

HEADERS:
- Add <algorithm> to CharacterLifeEvents.cpp for std::count_if, std::sort

Fixes 200+ cascading compilation errors across 11 files.

+++ src/game/character/CharacterRelationships.cpp
@@ -236,7 +236,7 @@ bool CharacterRelationshipsComponent::Deserialize(const std::string& json_str) {
     if (data.isMember("marriages") && data["marriages"].isArray()) {
         marriages.clear();
         const Json::Value& marriages_array = data["marriages"];
-        size_t max_marriages = std::min(marriages_array.size(),
+        size_t max_marriages = std::min(static_cast<size_t>(marriages_array.size()),
             static_cast<size_t>(game::core::serialization::MAX_MARRIAGES));
 
         for (Json::ArrayIndex i = 0; i < max_marriages; ++i) {
@@ -248,7 +248,7 @@ bool CharacterRelationshipsComponent::Deserialize(const std::string& json_str) {
     if (data.isMember("relationships") && data["relationships"].isArray()) {
         relationships.clear();
         const Json::Value& relationships_array = data["relationships"];
-        size_t max_relationships = std::min(relationships_array.size(),
+        size_t max_relationships = std::min(static_cast<size_t>(relationships_array.size()),
             static_cast<size_t>(game::core::serialization::MAX_RELATIONSHIPS));
 
         for (Json::ArrayIndex i = 0; i < max_relationships; ++i) {
@@ -262,7 +262,7 @@ bool CharacterRelationshipsComponent::Deserialize(const std::string& json_str) {
     if (data.isMember("children") && data["children"].isArray()) {
         children.clear();
         const Json::Value& children_array = data["children"];
-        size_t max_children = std::min(children_array.size(),
+        size_t max_children = std::min(static_cast<size_t>(children_array.size()),
             static_cast<size_t>(game::core::serialization::MAX_CHILDREN));
 
         for (Json::ArrayIndex i = 0; i < max_children; ++i) {
@@ -276,7 +276,7 @@ bool CharacterRelationshipsComponent::Deserialize(const std::string& json_str) {
     if (data.isMember("siblings") && data["siblings"].isArray()) {
         siblings.clear();
         const Json::Value& siblings_array = data["siblings"];
-        size_t max_siblings = std::min(siblings_array.size(),
+        size_t max_siblings = std::min(static_cast<size_t>(siblings_array.size()),
             static_cast<size_t>(game::core::serialization::MAX_CHILDREN));  // Reuse same limit
 
         for (Json::ArrayIndex i = 0; i < max_siblings; ++i) {
